+++ src/Imaging/Processing/Transforming/TransformationCreator.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Strider2038\ImgCache\Imaging\Transformation;
+namespace Strider2038\ImgCache\Imaging\Processing\Transforming;
 
 /**
  * @author Igor Lazarev <strider2038@rambler.ru>
