+++ src/Imaging/Processing/Transforming/TransformationFactoryInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Strider2038\ImgCache\Imaging\Transformation;
+namespace Strider2038\ImgCache\Imaging\Processing\Transforming;
 
 /**
  * @author Igor Lazarev <strider2038@rambler.ru>
