+++ tests/Unit/Imaging/Processing/Transformating/FlopTransformationTest.php
@@ -8,10 +8,10 @@
  * file that was distributed with this source code.
  */
 
-namespace Strider2038\ImgCache\Tests\Unit\Imaging\Transformation;
+namespace Strider2038\ImgCache\Tests\Unit\Imaging\Processing\Transforming;
 
 use Strider2038\ImgCache\Imaging\Processing\ImageTransformerInterface;
-use Strider2038\ImgCache\Imaging\Transformation\FlopTransformation;
+use Strider2038\ImgCache\Imaging\Processing\Transforming\FlopTransformation;
 use PHPUnit\Framework\TestCase;
 
 class FlopTransformationTest extends TestCase
