+++ tests/Unit/Imaging/Processing/Transformating/ResizeTransformationFactoryTest.php
@@ -9,16 +9,16 @@
  * file that was distributed with this source code.
  */
 
-namespace Strider2038\ImgCache\Tests\Unit\Imaging\Transformation;
+namespace Strider2038\ImgCache\Tests\Unit\Imaging\Processing\Transforming;
 
 use PHPUnit\Framework\TestCase;
 use Strider2038\ImgCache\Collection\StringList;
 use Strider2038\ImgCache\Enum\ResizeModeEnum;
 use Strider2038\ImgCache\Exception\InvalidRequestValueException;
 use Strider2038\ImgCache\Imaging\Parsing\StringParametersParserInterface;
-use Strider2038\ImgCache\Imaging\Transformation\ResizeParameters;
-use Strider2038\ImgCache\Imaging\Transformation\ResizeTransformation;
-use Strider2038\ImgCache\Imaging\Transformation\ResizeTransformationFactory;
+use Strider2038\ImgCache\Imaging\Processing\Transforming\ResizeParameters;
+use Strider2038\ImgCache\Imaging\Processing\Transforming\ResizeTransformation;
+use Strider2038\ImgCache\Imaging\Processing\Transforming\ResizeTransformationFactory;
 use Strider2038\ImgCache\Utility\EntityValidatorInterface;
 
 /**
