+++ tests/Unit/Imaging/Processing/Transformating/ResizeTransformationTest.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Strider2038\ImgCache\Tests\Unit\Imaging\Transformation;
+namespace Strider2038\ImgCache\Tests\Unit\Imaging\Processing\Transforming;
 
 use PHPUnit\Framework\TestCase;
 use Strider2038\ImgCache\Core\Streaming\StreamInterface;
@@ -19,8 +19,8 @@ use Strider2038\ImgCache\Imaging\Processing\ImageTransformerInterface;
 use Strider2038\ImgCache\Imaging\Processing\RectangleInterface;
 use Strider2038\ImgCache\Imaging\Processing\Size;
 use Strider2038\ImgCache\Imaging\Processing\SizeInterface;
-use Strider2038\ImgCache\Imaging\Transformation\ResizeParameters;
-use Strider2038\ImgCache\Imaging\Transformation\ResizeTransformation;
+use Strider2038\ImgCache\Imaging\Processing\Transforming\ResizeParameters;
+use Strider2038\ImgCache\Imaging\Processing\Transforming\ResizeTransformation;
 
 /**
  * @author Igor Lazarev <strider2038@rambler.ru>
@@ -115,6 +115,7 @@ class ResizeTransformationTest extends TestCase
 
             public function flip(): ImageTransformerInterface {}
             public function flop(): ImageTransformerInterface {}
+            public function rotate(float $degree): ImageTransformerInterface {}
             public function getImage(): Image {}
             public function setCompressionQuality(int $quality): ImageTransformerInterface {}
             public function writeToFile(string $filename): ImageTransformerInterface {}
