+++ tests/Unit/Imaging/Processing/Transformating/RotationParametersTest.php
@@ -8,9 +8,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Strider2038\ImgCache\Tests\Unit\Imaging\Transformation;
+namespace Strider2038\ImgCache\Tests\Unit\Imaging\Processing\Transforming;
 
-use Strider2038\ImgCache\Imaging\Transformation\RotationParameters;
+use Strider2038\ImgCache\Imaging\Processing\Transforming\RotationParameters;
 use PHPUnit\Framework\TestCase;
 use Strider2038\ImgCache\Utility\EntityValidator;
 use Strider2038\ImgCache\Utility\EntityValidatorInterface;
